fix training tensor parallel
