+++ recipes/wmt22_with_llama3.1/llama-instruct-inference.yaml
@@ -1,5 +1,5 @@
 # Model info
-model_path: "${EOLE_MODEL_DIR}/llama3.1-70b-instruct"
+model_path: "${EOLE_MODEL_DIR}/llama3.1-8b-instruct"
 
 # Inference
 seed: 42
@@ -12,8 +12,8 @@ batch_size: 512
 world_size: 2
 gpu_ranks: [0, 1]
 parallel_mode: "tensor_parallel"
-quant_layers: ['gate_up_proj', 'down_proj', 'up_proj', 'linear_values', 'linear_query', 'linear_keys', 'final_linear']
-quant_type: "bnb_NF4"
+#quant_layers: ['gate_up_proj', 'down_proj', 'up_proj', 'linear_values', 'linear_query', 'linear_keys', 'final_linear']
+#quant_type: "bnb_NF4"
 compute_dtype: fp16
 top_k: 0
 top_p: 0.0
